% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{create_gist}
\alias{create_gist}
\title{Create Github Gist files}
\usage{
create_gist(name, code, description)
}
\arguments{
\item{name}{The name of the Gist file.}

\item{code}{String with the code of the Gist.}

\item{description}{Code description.}
}
\value{
Message if successful.
}
\description{
The function \code{copycat_gists.create()} creates GitHub Gists.
}
