% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{delete_gist}
\alias{delete_gist}
\title{Delete Github Gist files}
\usage{
delete_gist(id)
}
\arguments{
\item{id}{The id of the Gist file.}
}
\value{
Message if file was successfully deleted.
}
\description{
Be careful: The function \code{delete_gist()} deletes GitHub Gists file.
}
