% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{get_gist}
\alias{get_gist}
\title{Copy Github Gist files}
\usage{
get_gist(filename, description = FALSE)
}
\arguments{
\item{filename}{The file name from one of your GitHub account (see gistfiles)}

\item{description}{The file name from one of your GitHub account (see gistfiles)}
}
\value{
Message if successful.
}
\description{
The function \code{get_gist()} copies GitHub Gists.
}
