% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIST-package.R
\docType{package}
\name{GIST-package}
\alias{GIST-package}
\alias{GIST}
\title{The 'GIST' package.}
\description{
Guiding-Image Spatial Transcriptomics
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
