% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIST.R
\name{GIST}
\alias{GIST}
\title{Guided Image Spatial Transcriptomics}
\usage{
GIST(
  st_expression,
  sig_mat,
  prior_values = NULL,
  prior_index = NULL,
  prior_lambda = 50,
  num_cores = 1,
  ...
)
}
\arguments{
\item{st_expression}{spatial transcriptomics expression matrix (genes by spot)}

\item{sig_mat}{signature matrix (genes by cell type)}

\item{prior_values}{vector of prior values}

\item{prior_index}{index in signature matrix of cell type which needs a prior on it}

\item{prior_lambda}{tunable hyperparameter \eqn{\lambda} for the prior beta distribution}

\item{...}{arguments to be passed to \code{rstan::sampling} (e.g. \code{chains, iter, init, verbose, refresh})}
}
\value{
An object of type dataframe with posterior mean estimates
}
\description{
Guided Image Spatial Transcriptomics
}
