% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{gist_addin}
\alias{gist_addin}
\title{Manage GitHub Gists via the gist addin}
\usage{
gist_addin()
}
\description{
The \code{gist_addin()} creates a shiny gadget
to manage your gists. Pick a gists, press the insert
button and the code will be insert into the current document at the
location of the cursor. Furthermore you can create and delete gists.
}
