% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{gist_create}
\alias{gist_create}
\title{Create a Gist}
\usage{
gist_create(name, code, description)
}
\arguments{
\item{name}{The name of the file.}

\item{code}{The code as character.}

\item{description}{The description of the code as character.}
}
\value{
Message if successful.
}
\description{
The function \code{gist_create()} creates a new Gist file.
}
