% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{gist_delete}
\alias{gist_delete}
\title{Delete a Gist}
\usage{
gist_delete(id)
}
\arguments{
\item{id}{The id of the Gist file (from gistFiles).}
}
\value{
Message if file was successfully deleted.
}
\description{
Be careful: The function \code{gist_delete()} deletes a Gist file.
}
