% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{gist_get}
\alias{gist_get}
\title{Get a Gist}
\usage{
gist_get(filename, description = FALSE, raw = FALSE)
}
\arguments{
\item{filename}{The file name of a gist file.}

\item{description}{If TRUE, the function returns the description of the Gist.}

\item{raw}{Get raw text string or copy it to clipboard.}
}
\value{
Message if successful.
}
\description{
The function \code{gist_get()} copies a Gist from Git Hub.
}
