% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{gistAddin}
\alias{gistAddin}
\title{The Gist Addin}
\usage{
gistAddin()
}
\description{
The \code{gistAddin()} creates a shiny app
to manage your gists. Pick a gist, press the copy or the insert
button and the code will be insert.
Furthermore you can create and delete gists as well.
}
