% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{gistfiles}
\alias{gistfiles}
\title{Explore GitHub Gist files}
\usage{
gistfiles()
}
\value{
Data frame with file names.
}
\description{
The function \code{copycat_gistfiles()} returns a list with GitHub Gists.
}
