% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load10xMatrix}
\alias{load10xMatrix}
\title{Read in MEX format data produced by Spaceranger
source:
https://support.10xgenomics.com/spatial-gene-expression/software/pipelines/latest/output/matrices}
\usage{
load10xMatrix(matrix_dir)
}
\arguments{
\item{matrix_dir}{Directory containing counts, barcode and features}
}
\value{
A dataframe of genes counts. Columns represet spots.
}
\description{
Read in MEX format data produced by Spaceranger
source:
https://support.10xgenomics.com/spatial-gene-expression/software/pipelines/latest/output/matrices
}
