% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load10xTP}
\alias{load10xTP}
\title{Read in the tissue positions file produced by Spaceranger
Ensures consistent columnn nomenclature}
\usage{
load10xTP(tp_file)
}
\arguments{
\item{tp_file}{Tissue positions file generated by Spaceranger}
}
\value{
a dataframe of tissue positions.
}
\description{
Read in the tissue positions file produced by Spaceranger
Ensures consistent columnn nomenclature
}
