% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_signature_matrix}
\alias{make_signature_matrix}
\title{Make signature matrix from
imputed, filtered, normalized
single cell counts}
\usage{
make_signature_matrix(sc_counts, sc_labels, save_fname = NULL)
}
\arguments{
\item{sc_counts}{A matrix/dataframe of single cell counts, with cells as columns and genes as rows}

\item{sc_labels}{A matrix/dataframe of cell type labels, two cols, one for cell and one for cell type}

\item{save_fname}{Optionally save the signature matrix}
}
\value{
a matrix of gene expression means with genes as rows and cell types as columns
}
\description{
Make signature matrix from
imputed, filtered, normalized
single cell counts
}
