% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{merge_with_locations}
\alias{merge_with_locations}
\title{Merge dataframe with tissue positions
internal function for plotting data
tp should be a dataframe with pxl_col, pxl_row}
\usage{
merge_with_locations(.data, tp)
}
\arguments{
\item{.data}{a dataframe with rownames as barcodes}

\item{tp}{a dataframe of tissue positions}
}
\value{
a merged dataframe in long format
}
\description{
Merge dataframe with tissue positions
internal function for plotting data
tp should be a dataframe with pxl_col, pxl_row
}
