% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize.decon}
\alias{normalize.decon}
\title{Normalize a genes by cell matrix, and
optionally perform some QC}
\usage{
normalize.decon(
  data,
  doQC = FALSE,
  method = c("scale", "quantile", "scale-quantile", "sct"),
  only_hvg = T
)
}
\arguments{
\item{data}{A genes x cell matrix}

\item{doQC}{Should QC be performed: min.cells = 10\%, min.features = 500}

\item{method}{Choose a method from 'scale', 'quantile', 'scale-quantile', 'sct'}

\item{only_hvg}{If using method 'sct', return only highly variable genes after normalization}
}
\value{
a matrix of normalized gene counts
}
\description{
Normalize a genes by cell matrix, and
optionally perform some QC
}
