% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalize.quantiles2}
\alias{normalize.quantiles2}
\title{Normalize quantiles wrapper
to handle loss of rownames
and colnames}
\usage{
normalize.quantiles2(counts_mat)
}
\arguments{
\item{counts_mat}{Matrix of counts to be normalized}
}
\value{
a quantile normalzed of gene counts
}
\description{
Normalize quantiles wrapper
to handle loss of rownames
and colnames
}
