% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_spots}
\alias{plot_spots}
\title{Visualization of spots with proportions of cell types}
\usage{
plot_spots(
  df,
  tp,
  trans = "identity",
  legend.label = "Prop.",
  ncols = 3,
  joint_scale = FALSE,
  size_m = 1,
  my_theme = NULL
)
}
\arguments{
\item{df}{Spot by value matrix, spots are rows, values are cols}

\item{tp}{Matrix of spot locations, should contain
\code{pxl_col}, \code{pxl_row} indication column and row location of pixel}

\item{trans}{Any transformation that needs to be done to values - log10 etc.}

\item{legend.label}{Label for the value legend (default: "Prop.")}

\item{ncols}{number of columns}

\item{joint_scale}{Plot common legend scale (default: FALSE)}

\item{size_m}{Size of the spots to be plotted (default: 1)}

\item{my_theme}{gg_theme to be passed to ggplot}
}
\value{
a visualization of supplied metrics according to tissues positions
}
\description{
Visualization of spots with proportions of cell types
}
