% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_spots0}
\alias{plot_spots0}
\title{Visualization of spots expression
across tissue for a single variable}
\usage{
plot_spots0(df, tp, trans = "identity", legend.label = "Prop.", size_m = 4)
}
\arguments{
\item{df}{Spot by value matrix, spots are rows, values are cols}

\item{tp}{Matrix of spot locations, should contain
pxl_col, pxl_row indicating column and row location of pixel}

\item{trans}{Any transformation that needs to be done to values - log10 etc. (default: "identity")}

\item{legend.label}{Label for the value legend (default: "Propr.")}

\item{size_m}{size of the spots to be plotted (default: 4)}
}
\value{
plot showing spots according to tissue positions
}
\description{
Visualization of spots expression
across tissue for a single variable
}
