% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{create_gist}
\alias{create_gist}
\title{Create Gist file}
\usage{
create_gist(name, code, description)
}
\arguments{
\item{name}{The name of the file.}

\item{code}{The code as character.}

\item{description}{The description of the code as character.}
}
\value{
Message if successful.
}
\description{
The function \code{create_gist()} creates a new Gist file.
}
