% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{create_gist}
\alias{create_gist}
\title{Create Github Gist files}
\usage{
create_gist(name, code, description)
}
\arguments{
\item{name}{The name of the file.}

\item{code}{String with the code of the Gist.}

\item{description}{String with the code description.}
}
\value{
Message if successful.
}
\description{
The function \code{create_gist()} creates a new GitHub Gist file.
}
