% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{delete_gist}
\alias{delete_gist}
\title{Delete Github Gist files}
\usage{
delete_gist(id)
}
\arguments{
\item{id}{The id of the Gist file (see copycat_gistfiles).}
}
\value{
Message if successful.
}
\description{
The function \code{copycat_gists.del()} deletes GitHub Gists.
}
