% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{get_gist}
\alias{get_gist}
\title{Get Gist file}
\usage{
get_gist(filename, description = FALSE, raw = FALSE)
}
\arguments{
\item{filename}{The file name of a gist file.}

\item{description}{If TRUE, the function returns the description of the Gist.}

\item{raw}{Get raw text string or copy}
}
\value{
Message if successful.
}
\description{
The function \code{get_gist()} copies a Gist file.
}
