% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{gistFiles}
\alias{gistFiles}
\title{List All Gist Files}
\usage{
gistFiles(arrange = TRUE)
}
\arguments{
\item{arrange}{To arrange the files alphabetically, set to TRUE.}
}
\value{
Data frame with file names and id.
}
\description{
The function \code{gistFiles()} returns all Gist files
from a given Git Hub account.
}
