% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gists.R
\name{gistfiles}
\alias{gistfiles}
\title{List all Gist files}
\usage{
gistfiles(arrange = TRUE)
}
\arguments{
\item{arrange}{If the files should be arranged alphabetically.}
}
\value{
Data frame with file names and id.
}
\description{
The function \code{gistfiles()} returns all Gists file.
}
